refactor(vize): split commands/lint into cohesive modules
